Update AsyncAwait README for clarity and correctness
In this commit, we made modifications in the Lesson.AsyncAwait/README.md file. The first part of the document was updated with clearer terms and descriptions. A correction was also applied to the usage of the term 'асинхронномв' which was corrected to 'асинхронному', and additional information was attached for viewing the original code and state machine on sharplab.io.